per review: refactored ActorRole to its own file, used in vehicle_index
