+++ smarts/core/agent_manager.py
@@ -24,13 +24,14 @@ from typing import Any, Dict, Optional, Set, Tuple, Union
 import cloudpickle
 
 from envision.types import format_actor_id
+from smarts.core.actor_role import ActorRole
 from smarts.core.agent_interface import AgentInterface
 from smarts.core.bubble_manager import BubbleManager
 from smarts.core.data_model import SocialAgent
 from smarts.core.plan import Plan
 from smarts.core.sensors import Observation, Sensors
 from smarts.core.utils.id import SocialAgentId
-from smarts.core.vehicle import ActorRole, VehicleState
+from smarts.core.vehicle import VehicleState
 from smarts.zoo.registry import make as make_social_agent
 
 
@@ -481,7 +482,7 @@ class AgentManager:
             agent_model.initial_speed,
             boid=boid,
         )
-        role = ActorRole.SocialAgent
+        role = ActorRole.EgoAgent if trainable else ActorRole.SocialAgent
         for provider in sim.providers:
             if agent_interface.action_space not in provider.action_spaces:
                 continue
